Rename package to point-annotator
